Fix to use int values for temperature.

+++ thermostat/manager.h
@@ -13,7 +13,7 @@ class QNetworkReply;
 
 struct TimelineSlotsData {
     int  onOff;
-    float tempSP;
+    int tempSP;
 };
